requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
langchain==0.2.6
langchain-openai==0.1.11
faiss-cpu==1.8.0
tiktoken==0.7.0
pypdf==4.3.1
